% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_so.R
\name{find_so}
\alias{find_so}
\title{Find shared object.}
\usage{
find_so(package_name)
}
\arguments{
\item{package_name}{A single string.  The package to search for shared
objects.}
}
\value{
An array of strings containing the shared objects found, including 
  paths.  If no shared objects are found or the package is not installed
  the function records an empty array.
}
\description{
\code{find_so} returns the shared objects found within the lib directory
of an R package.
}
\details{
This is a function that searches a specified and installed R package's 
lib directory for shared objects, and returns any shared objects found.
}
\examples{
find_so('utils')
}
