% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sysdep.R
\name{get_sysdep}
\alias{get_sysdep}
\title{Get system dependencies from online database.}
\usage{
get_sysdep(r_pkg_name, r_pkg_version = "", os_name, os_detail, r_version,
  sys_pkg_system, arch, url = "http://10.0.1.70/index.php")
}
\arguments{
\item{r_pkg_name}{A single string.  The package to search for in the 
dependency database.}

\item{r_pkg_version}{A single string.  The package version to search for 
in the dependency database. (default is the most current version)}

\item{os_name}{A single string.  The operating system name to search for 
in the dependency database.  (default is the current operating system)}

\item{os_detail}{A single string.  The operating system version name to 
search for in the dependency database.  (default is the current 
operating system version)}

\item{r_version}{A single string.  The R version to search for in the
dependency database.  (default is the current R version)}

\item{sys_pkg_system}{A single string.  The package system to search for 
in the dependency database.  (default is based on current operating 
system)}

\item{arch}{A single string.  The processor architecture to search for 
in the dependency database.  (default is the current processor 
architecture)}

\item{url}{A single string.  The URL for the online database.}
}
\value{
A data frame containing:
 
r_pkg_name:  R package name.

r_pkg_version:  R package version.

os_name:  Operating system.

os_detail:  Operating system version.

r_version:  R version.

sys_pkg_system:  Package system, e.g. homebrew, dpkg, and rpm.

arch:  Processor architecture of system.

url:  URL of online database.

When no dependencies are found, NULL is returned.
}
\description{
\code{get_sysdep} returns R package system dependencies using an online
database.
}
\details{
This is a function that searches an online database for R system 
dependencies.  This database includes Makevar specifications.
}
\examples{
\dontrun{
get_sysdep('rgdal')
}
}
