% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dep_bin.R
\name{pkg_dep_bin}
\alias{pkg_dep_bin}
\title{Creates a list of run-time system dependencies for installed packages.}
\usage{
pkg_dep_bin(pkg_names, verbose = TRUE)
}
\arguments{
\item{pkg_names}{A string array of installed packages.}

\item{verbose}{A logical scalar. Should additional output be provided, 
(default is TRUE).}
}
\value{
A data frame containing:
 
r_package:  R package name.

r_package_version:  R package version.

operating_system:  Current operating system.

shared_object:  Shared object.

shared_object_exists:  Logical, does the referred shared object exist.

package_system:  Package system, e.g. homebrew, dpkg, and rpm.

package_version:  Version of the package.

package_name:  System package name.

When no dependencies are found, NULL is returned.
}
\description{
\code{pkg_dep_bin} returns a list of run-time system dependencies based 
on shared objects.  System dependencies that cannot be determined through
shared objects are not returned.
}
\examples{
pkg_dep_bin(c('stats','utils'))
}
